--- src/Components/Agenda/AgendaLeg/index.js
@@ -1,1 +1,0 @@
-export { default } from './AgendaLeg';
