+++ src/Components/Agenda/AgendaLegFormEdit/index.js
@@ -1,0 +1,1 @@
+export { default } from './AgendaLegFormEdit';
